+++ mulog-core/dev/your_ns.clj
@@ -74,6 +74,7 @@
 
 
 
+
 (comment
 
   (def stop-all
@@ -87,5 +88,90 @@
   (μ/log ::hello :to "New World!")
 
   (stop-all)
+
+  )
+
+
+
+
+
+
+(comment
+
+  (require '[com.brunobonacci.mulog.buffer :as rb]
+           '[clojure.pprint :refer [pprint]])
+
+
+  (deftype MyCustomPublisher
+      [config buffer]
+
+
+    com.brunobonacci.mulog.publisher.PPublisher
+    (agent-buffer [_]
+      buffer)
+
+
+    (publish-delay [_]
+      500)
+
+
+    (publish [_ buffer]
+      ;; check our printer option
+      (let [printer (if (:pretty-print config) pprint prn)]
+        ;; items are pairs [offset <item>]
+        (doseq [item (map second (rb/items buffer))]
+          ;; print the item
+          (printer item)))
+      ;; return the buffer minus the published elements
+      (rb/clear buffer)))
+
+
+  (defn my-custom-publisher
+    [config]
+    (MyCustomPublisher. config (rb/agent-buffer 10000)))
+
+
+  (defn- pprint-str
+    [v]
+    (with-out-str
+      (pprint v)))
+
+
+  (deftype MyCustomPublisher
+      [config buffer ^java.io.Writer filewriter]
+
+
+    com.brunobonacci.mulog.publisher.PPublisher
+    (agent-buffer [_]
+      buffer)
+
+
+    (publish-delay [_]
+      500)
+
+
+    (publish [_ buffer]
+      ;;    check our printer option
+      (let [printer (if (:pretty-print config) pprint-str prn-str)
+            ;; take at most `:max-items` items
+            items (take (:max-items config) (rb/items buffer))
+            ;; save the offset of the last items
+            last-offset (-> items last first)]
+        ;; write the items to the file
+        (doseq [item (map second items)]
+          ;; print the item
+          (.write filewriter (printer item)))
+        ;; flush the buffer
+        (.flush filewriter)
+        ;; return the buffer minus the published elements
+        (rb/dequeue buffer last-offset))))
+
+
+  (defn my-custom-publisher
+    [{:keys [filename] :as config}]
+    (let [config (merge {:pretty-print false :max-items 1000} config)]
+      (MyCustomPublisher. config (rb/agent-buffer 10000)
+                          (io/writer (io/file filename) :append true))))
+
 
   )